corrected variant
